# docs/swagger.yaml
basePath: /
definitions:
...
  shopping_api.CancelOrderRequest:
    properties:
      orderId:
        type: integer
    type: object
...
  user_api.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Basket service api provides category-product-order.
  title: Gin Basket Service API main
  version: "1.0"
paths:
...
  /order/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create Order
      tags:
      - Order
  /order/cancel:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order Id
        in: body
        name: cancelOrderRequest
        schema:
          $ref: '#/definitions/shopping_api.CancelOrderRequest'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Cancel Order
      tags:
      - Order
  /order/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get All Order
      tags:
      - Order
...
  /user/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: User Info
        in: body
        name: loginRequest
        schema:
          $ref: '#/definitions/user_api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add user
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: User Info
        in: body
        name: loginRequest
        schema:
          $ref: '#/definitions/user_api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - User
schemes:
- http
swagger: "2.0"